+++ integrationTests/chainSimulator/esdtPrefixBridge/executeAndBurn_test.go
@@ -1,4 +1,4 @@
-package bridge
+package esdtPrefixBridge
 
 import (
 	"encoding/hex"
