+++ integrationTests/chainSimulator/esdtPrefixBridge/testdata/enshrine-esdt-safe.wasm
