+++ integrationTests/chainSimulator/esdtPrefixBridge/testdata/fee-market.wasm
