+++ integrationTests/chainSimulator/esdtPrefixBridge/testdata/hello.wasm
