+++ integrationTests/chainSimulator/esdtPrefixBridge/testdata/simple-enshrine-esdt-safe.wasm
