+++ integrationTests/chainSimulator/esdtPrefixBridge/testdata/simple-esdt-test.wasm
